moved groupmenu styles to external file
